Move `VerifySession` to command package.
