+++ sessions/src/main/proto/spine_examples/pingh/sessions/session_verification_commands.proto
@@ -31,7 +31,7 @@ package spine_examples.pingh.sessions;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.pingh.spine.io";
-option java_package = "io.spine.examples.pingh.sessions.event";
+option java_package = "io.spine.examples.pingh.sessions.command";
 option java_outer_classname = "SessionVerificationCommandProto";
 option java_multiple_files = true;
